+++ neosvr_headless_webui/dashboard.py
@@ -1,0 +1,16 @@
+from flask import Blueprint, current_app, g, render_template
+from .auth import login_required
+
+from rpyc import connect
+
+bp = Blueprint("dashboard", __name__, url_prefix="/dashboard")
+
+@bp.route("")
+@login_required
+def dashboard():
+    conn = connect(
+        current_app.config["MANAGER_HOST"], current_app.config["MANAGER_PORT"]
+    )
+    g.clients = conn.root.list_headless_clients()
+    g.status = conn.root.get_manager_status()
+    return render_template("dashboard.html")